WebAPI/requirements.txt
Django
django-cors-headers
djangorestframework
pytz
pandas
numpy
nltk
sklearn
tensorflow
keras
os
pickle
json